+++ packages/workspace-tools/CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.177.2 (2024-09-08)
 
-
 ### Bug Fixes
 
-- **workspace-tools:** Added console logging to error reporting ([4878661d](https://github.com/storm-software/storm-ops/commit/4878661d))
+- **workspace-tools:** Added console logging to error reporting
+  ([4878661d](https://github.com/storm-software/storm-ops/commit/4878661d))
 
 ## 1.177.1 (2024-09-08)